+++ Models/Accounts/Request/AuthenticateRequest.cs
@@ -1,11 +1,14 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace WebAlbum.Models.Accounts
+namespace WebAlbum.Models.Accounts.Request
 {
-    public class ForgotPasswordRequest
+    public class AuthenticateRequest
     {
         [Required]
         [EmailAddress]
         public string Email { get; set; }
+
+        [Required]
+        public string Password { get; set; }
     }
 }
