+++ Models/Accounts/Request/RegisterRequest.cs
@@ -1,17 +1,11 @@
-﻿using WebAlbum.Entities;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace WebAlbum.Models.Accounts
+namespace WebAlbum.Models.Accounts.Request
 {
-    public class CreateRequest
+    public class RegisterRequest
     {
-        
         [Required]
         public string FirstName { get; set; }
-
-        [Required]
-        [EnumDataType(typeof(Role))]
-        public string Role { get; set; }
 
         [Required]
         [EmailAddress]
@@ -24,5 +18,8 @@ namespace WebAlbum.Models.Accounts
         [Required]
         [Compare("Password")]
         public string ConfirmPassword { get; set; }
+
+        [Range(typeof(bool), "true", "true")]
+        public bool AcceptTerms { get; set; }
     }
 }
