+++ Models/Accounts/Request/UpdateRequest.cs
@@ -1,7 +1,7 @@
 ﻿using WebAlbum.Entities;
 using System.ComponentModel.DataAnnotations;
 
-namespace WebAlbum.Models.Accounts
+namespace WebAlbum.Models.Accounts.Request
 {
     public class UpdateRequest
     {
