+++ Models/Accounts/Request/ValidateResetTokenRequest.cs
@@ -1,8 +1,8 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace WebAlbum.Models.Accounts
+namespace WebAlbum.Models.Accounts.Request
 {
-    public class VerifyEmailRequest
+    public class ValidateResetTokenRequest
     {
         [Required]
         public string Token { get; set; }
