+++ Models/Accounts/Response/AccountResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace WebAlbum.Models.Accounts
+﻿namespace WebAlbum.Models.Accounts.Response
 {
     public class AccountResponse
     {
