+++ Models/Images/ImagesResponse.cs
@@ -1,0 +1,10 @@
+﻿namespace WebAlbum.Models.Images
+{
+    public class ImagesResponse
+    {
+        public int Id { get; set; }
+        public string Title { get; set; }
+        public string Description { get; set; }
+        public int? FolderId { get; set; }
+    }
+}
